Rename to fused_attn_fp16_bf16_max_seqlen_512 for the better readability
