+++ transformer_engine/common/fused_attn/fused_attn.cpp
@@ -7,8 +7,8 @@
 #include "transformer_engine/fused_attn.h"
 #include "../common.h"
 #include "utils.h"
+#include "fused_attn_fp16_bf16_max_seqlen_512.h"
 #include "fused_attn_fp8.h"
-#include "fused_attn_max_512.h"
 
 // NVTE fused attention FWD FP8 with packed QKV
 void nvte_fused_attn_fwd_qkvpacked(
